firewall/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyYAML==6.0.2
httpx==0.27.2
